+++ ydb/core/kqp/common/kqp_user_request_context.h
@@ -6,7 +6,7 @@
 
 #include <ydb/core/resource_pools/resource_pool_settings.h>
 #include <ydb/library/actors/core/actorid.h>
-#include <yql/essentials/core/cbo/cbo_optimizer_new.h>
+#include <ydb/core/kqp/opt/cbo/cbo_optimizer_new.h>
 
 namespace NYql::NPq::NProto {
     class StreamingDisposition;
@@ -28,7 +28,7 @@ namespace NKikimr::NKqp {
         TString CheckpointId;
         TString StreamingQueryPath;
         std::shared_ptr<NYql::NPq::NProto::StreamingDisposition> StreamingDisposition;
-        std::shared_ptr<NYql::TOptimizerHints> Hints{};
+        std::shared_ptr<NKikimr::NKqp::TOptimizerHints> Hints{};
         TMaybe<ui32> CostBasedOptimizationLevelOverride;
 
         TUserRequestContext() = default;
