+++ ydb/library/yql/dq/opt/dq_opt_join_cost_based.h
@@ -26,8 +26,7 @@ NYql::NNodes::TExprBase DqOptimizeEquiJoinWithCosts(
     const TProviderCollectFunction& providerCollect,
     const TOptimizerHints& hints = {},
     bool enableShuffleElimination = false,
-    NYql::TShufflingOrderingsByJoinLabels* shufflingOrderingsByJoinLabels = nullptr,
-    TOptimizerTrueCardinalitiesHints* costBasedHints = nullptr
+    NYql::TShufflingOrderingsByJoinLabels* shufflingOrderingsByJoinLabels = nullptr
 );
 
 NYql::NNodes::TExprBase DqOptimizeEquiJoinWithCosts(
@@ -40,8 +39,7 @@ NYql::NNodes::TExprBase DqOptimizeEquiJoinWithCosts(
     int& equiJoinCounter,
     const TOptimizerHints& hints = {},
     bool enableShuffleElimination = false,
-    NYql::TShufflingOrderingsByJoinLabels* shufflingOrderingsByJoinLabels = nullptr,
-    TOptimizerTrueCardinalitiesHints* costBasedHints = nullptr
+    NYql::TShufflingOrderingsByJoinLabels* shufflingOrderingsByJoinLabels = nullptr
 );
 
 void CollectInterestingOrderingsFromJoinTree(
