+++ yql/essentials/sql/pg/optimizer.cpp
@@ -697,9 +697,8 @@ public:
 
     std::shared_ptr<TJoinOptimizerNode> JoinSearch(
         const std::shared_ptr<TJoinOptimizerNode>& joinTree,
-        const TOptimizerHints& hints = {},
-        TOptimizerTrueCardinalitiesHints* costBasedHints = nullptr) override {
-        Y_UNUSED(hints, costBasedHints);
+        const TOptimizerHints& hints = {}) override {
+        Y_UNUSED(hints);
         return TPgOptimizerImpl(joinTree, Ctx_, Log_).Do();
     }
